je commit les premier test de classe pour les recherche et crud

+++ src/model/BusinessEntity.java
@@ -3,17 +3,17 @@ package model;
 public class BusinessEntity {
     private String reference; // primaryKey
     private String name;
-    private int vatNumber;
+    private Long vatNumber;
     private String statutReference; // foreignKey
 
-    public BusinessEntity(String reference, String name, int vatNumber, String statutReference){
+    public BusinessEntity(String reference, String name, Long vatNumber, String statutReference){
         this.reference = reference;
         this.name = name;
         setVatNumber(vatNumber);
         this.statutReference = statutReference;
     }
 
-    public void setVatNumber(int vatNumber) {
+    public void setVatNumber(Long vatNumber) {
         //
     }
 }
